+++ src/activityDashboard.ts
@@ -13,4 +13,4 @@ export type { TimelineBin, TimelineLane, TimelineModel } from "./activityDashboa
 export { buildTimeline } from "./activityDashboard/timeline.js";
 export { collectProjectDiff, collectProjectGit, resetGitStatusCache } from "./activityDashboard/git.js";
 export { collectActivityDashboard } from "./activityDashboard/collect.js";
-export { renderActivityBatchPage, renderActivityDashboardPage, renderProjectDiffFragment } from "./activityDashboard/page.js";
+export { renderActivityBatchFragment, renderActivityBatchPage, renderActivityDashboardPage, renderProjectDiffFragment } from "./activityDashboard/page.js";